+++ library/includes/classes/class-sanitize.php
@@ -8,7 +8,7 @@
  * @subpackage  Core
  *
  * @since    2.5.0
- * @version  2.5.6
+ * @version  2.6.1
  *
  * Contents:
  *
@@ -90,7 +90,7 @@ final class {%= prefix_class %}_Library_Sanitize {
 		 * Returns an array of values.
 		 *
 		 * @since    2.5.0
-		 * @version  2.5.4
+		 * @version  2.6.1
 		 *
 		 * @param  mixed $value
 		 * @param  array $choices
@@ -132,7 +132,7 @@ final class {%= prefix_class %}_Library_Sanitize {
 
 					$value[ $key ] = esc_attr( $single_value );
 
-				} // /foreach
+				}
 
 
 			// Output